+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 2.3.3
 - Maximum relative sapwood growth rates effective
+- New option 'summary.freq' in plot.spwb and plot.growth
 
 # Version 2.3.2
 - Parameter 'ParticleDensity' eliminated, as it is now calculated from 'LeafDensity', 'WoodDensity' and 'r635'